Fixes #167
